# pulse-query-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hospital Analytics Platform",
    "description": "Backend API for the Hospital Analytics Platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/settings": {
      "get": {
        "tags": ["admin"],
        "summary": "Read Admin Settings",
        "description": "Get system-wide admin settings like API keys and visible models.",
        "operationId": "read_admin_settings_api_v1_admin_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": ["admin"],
        "summary": "Write Admin Settings",
        "description": "Update system-wide admin settings like API keys and visible models.",
        "operationId": "write_admin_settings_api_v1_admin_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminSettingsResponse": {
        "properties": {
          "api_keys": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Api Keys"
          },
          "visible_models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Visible Models"
          }
        },
        "type": "object",
        "required": ["api_keys", "visible_models"],
        "title": "AdminSettingsResponse"
      },
      "AdminSettingsUpdateRequest": {
        "properties": {
          "api_keys": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Api Keys"
          },
          "visible_models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Visible Models"
          }
        },
        "type": "object",
        "required": ["api_keys", "visible_models"],
        "title": "AdminSettingsUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}